Added stream response handler class and updated method signatures

+++ library/src/main/java/net/callumtaylor/asynchttp/response/BitmapResponseHandler.java
@@ -13,7 +13,7 @@ public class BitmapResponseHandler extends ResponseHandler<Bitmap>
 	private ByteArrayOutputStream byteBuffer;
 	private Bitmap bitmap;
 
-	@Override public void onByteChunkReceived(byte[] chunk, int chunkLength, long totalProcessed, long totalLength)
+	@Override public void onByteChunkReceived(byte[] chunk, long chunkLength, long totalProcessed, long totalLength)
 	{
 		if (byteBuffer == null)
 		{
@@ -23,7 +23,7 @@ public class BitmapResponseHandler extends ResponseHandler<Bitmap>
 
 		if (chunk != null)
 		{
-			byteBuffer.write(chunk, 0, chunkLength);
+			byteBuffer.write(chunk, 0, (int)chunkLength);
 		}
 	}
