+++ library/src/main/java/net/callumtaylor/asynchttp/response/StringResponseHandler.java
@@ -8,7 +8,7 @@ public class StringResponseHandler extends ResponseHandler<String>
 	private StringBuffer stringBuffer;
 	private String content;
 
-	@Override public void onByteChunkReceived(byte[] chunk, int chunkLength, long totalProcessed, long totalLength)
+	@Override public void onByteChunkReceived(byte[] chunk, long chunkLength, long totalProcessed, long totalLength)
 	{
 		if (stringBuffer == null)
 		{
@@ -20,7 +20,7 @@ public class StringResponseHandler extends ResponseHandler<String>
 		{
 			try
 			{
-				stringBuffer.append(new String(chunk, 0, chunkLength, "UTF-8"));
+				stringBuffer.append(new String(chunk, 0, (int)chunkLength, "UTF-8"));
 			}
 			catch (Exception e)
 			{
